refactor: split main-menu styles to separated file
