+++ service_worker.js
@@ -20,7 +20,8 @@ const assets = [
   "/src/config/routes.js",
   "/src/components/circle-viewer/index.js",
   "/src/components/circle-viewer/styles.js",
-  "/src/components/main-menu.js",
+  "/src/components/main-menu/index.js",
+  "/src/components/main-menu/styles.js",
   "/src/view/index.js",
   "/src/view/errors/not-found.js",
   "/src/view/pages/about.js",
